feat: improve extensibility of last discussion widget
